+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-## Unreleased
+## 0.3.0
+- Add basic style
 - Support database as schema source.
 
 ## 0.2.0